chore(teldrive): organize code
